requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0 